Write test case for `currentStateName` bug.
This test fails because `gotoState` closes over `currentStateName`. This results in the each instance identifying its current state as the last state that was transitioned into and doing the wrong cleanup to itself when transitioning to a new state.

For example, in this test case, the first instance maintains the `testProperty1` because, internally, it identifies its state as `TestState2` and does the cleanup for that state which does not know about `testProperty1`.